+++ client/web/src/global/CoolCodeIntel.module.scss
