+++ client/web/src/global/CoolCodeIntelQueries.tsx
